Adding Loading/Error experience for signup page
