+++ my-app/src/App.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { Provider, useSelector } from "react-redux";
 import "./App.css";
-import { AuthPageContainer } from "./Pages/AuthPageConainter";
 import { ChatPage } from "./Pages/ChatPage";
 import { store } from "./Store/store";
 import { BrowserRouter, Switch, Route, Link } from "react-router-dom";
@@ -9,6 +8,7 @@ import { selectIsLoggedin, selectUserProfile } from "./Store/loginSlice";
 import io, { Socket } from 'socket.io-client';
 import { DefaultEventsMap } from "socket.io-client/build/typed-events";
 import SocketConnection from "./DataLayer/Socket/SocketConnection";
+import { AuthPage } from "./Pages/AuthPage";
 function App() {
   return (
     <Provider store={store}>
@@ -21,7 +21,7 @@ const EntryPoint: React.FC<{}> = () => {
   const selectedIsLoggedin = useSelector(selectIsLoggedin);
 
   if (!selectedProfile || !selectedIsLoggedin) {
-    return <AuthPageContainer />;
+    return <AuthPage />;
   }
   return (
     <BrowserRouter>
