+++ my-app/src/Models/Status.type.ts
@@ -1,0 +1,6 @@
+export enum Status {
+    Running,
+    Failed,
+    Completed,
+    NotStarted,
+  }
