--- my-app/src/Pages/AuthPageConainter.tsx
@@ -1,13 +1,0 @@
-import {useState} from 'react';
-import * as React from 'react';
-import {SignupPageFC} from './SignupPageFC';
-import { SigninPageFC } from './SigninPageFC';
-
-export const AuthPageContainer: React.FC<{}> = (): JSX.Element => {
-  const [isSignIn, setIsSignIn] = useState(false);
-  if (!isSignIn) {
-    return <SignupPageFC setIsSignIn={setIsSignIn}/>;
-  } else {
-    return <SigninPageFC setIsSignIn={setIsSignIn}/>;
-  }
-};
