--- my-app/src/Pages/SignupPage.css
@@ -1,11 +1,0 @@
-.signUpCard {
-    display: flex;
-    flex-direction: column;
-    width: 500px;
-    height: 600px;
-    margin: auto;
-}
-
-.signUpFields {
-    width: 100%;
-}
